refactor(ast): Extract error to root of crate
