+++ common/ast/src/error.rs
@@ -39,10 +39,11 @@ pub struct Error<'a> {
     errors: Vec<ErrorKind>,
     /// The backtrace stack of the error.
     contexts: Vec<(Token<'a>, &'static str)>,
-    /// The exrta backtrace of error in optional branches.
+    /// The extra backtrace of error in optional branches.
     backtrace: &'a Backtrace<'a>,
 }
 
+/// ErrorKind is the error type returned from parser.
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum ErrorKind {
     /// Error generated by `match_token` function
@@ -171,6 +172,7 @@ impl From<ParseIntError> for ErrorKind {
     }
 }
 
+/// DisplayError is used to display errors in user-friendly way.
 pub trait DisplayError {
     type Message;
